Fix bug with the "delete all" feature
In the "delete all" function, in order to remove the workouts from the DOM, I was selecting all the <li> elements of the document, since the workouts <li> were the only ones. But now that I added the filter pane, there are more <li> on the page and these should not be deleted. Lesson learned: it's better to be precise regarding selections since the very beginning, you never know what features you might want to add next.